Pjdeasy86 patch 1 (#17)
* Updated save_as_csv add on

PLEASE CHECK.
This adds in a function called save_as_csv file that can can be used to convert an object to a csv entry. The function receives the file name and object to be converted. If the file does not exist, it creates a file (with a header). If the file exists, it appends a new line with the object in the CSV. It has a method in each of the classes with the same name, which will return a string with the data already structured as a CSV.

When data is collected from a repositories, is is added to a CSV called repositories.csv . When it collects the pull requests of a repository, it stores them in a file named after the owner and the name of repository(repos/owner-repo.csv) when you collect data from users, It adds it to a CSV called users.csv.

* Update Deasy Data Pull.2

* Update gitdata.py

This adds in the save_as_csv requirements.

---------

Co-authored-by: j-tenny <107655580+j-tenny@users.noreply.github.com>